requirements.txt
bert_score==0.3.13
fire==0.7.0
fitz==0.0.1.dev2
langchain==0.3.19
langchain_anthropic==0.3.8
langchain_community==0.3.18
langchain_core==0.3.40
langchain_openai==0.3.7
numpy==2.2.3
pdf2image==1.17.0
Pillow==11.1.0
pydantic==2.10.6
PyPDF2==3.0.1
Requests==2.32.3
retry==0.9.2
rouge==1.0.1
skimage==0.0
tqdm==4.67.1
